+++ js/ucb-hero-unit.js
@@ -19,7 +19,7 @@
                         playerVars: { autoplay: 1, controls: 0, mute: 1, disablekb: 1, rel: 0, playsinline: 1 },
                         events: {
                             onReady: function(event){
-                                enableVideoHeroAutoresize(videoWrapperElement, videoPlayerWrapperElement, document.getElementById(videoPlayerElementId), 800, 450);
+                                // enableVideoHeroAutoresize(videoWrapperElement, videoPlayerWrapperElement, document.getElementById(videoPlayerElementId), 800, 450);
                                 videoWrapperElement.removeAttribute('hidden');
                             },
                             onStateChange: function(event) {
@@ -74,14 +74,14 @@
                     background: true,
                     muted: true
                 });
-                // Video width and height come back from the Vimeo player API as Promises
-                // In case of error, it will default to 800x450 which is fine for 16:9 videos
-                videoPlayer.getVideoWidth().then(
-                    (value) => { videoWidth = value; enableVideoHeroAutoresize(videoWrapperElement, videoPlayerWrapperElement, videoPlayer.element, videoWidth, videoHeight); },
-                    (error) => { videoWidth = 800; enableVideoHeroAutoresize(videoWrapperElement, videoPlayerWrapperElement, videoPlayer.element, videoWidth, videoHeight); });
-                videoPlayer.getVideoHeight().then(
-                    (value) => { videoHeight = value; enableVideoHeroAutoresize(videoWrapperElement, videoPlayerWrapperElement, videoPlayer.element, videoWidth, videoHeight); },
-                    (error) => { videoHeight = 450; enableVideoHeroAutoresize(videoWrapperElement, videoPlayerWrapperElement, videoPlayer.element, videoWidth, videoHeight); });
+                // // Video width and height come back from the Vimeo player API as Promises
+                // // In case of error, it will default to 800x450 which is fine for 16:9 videos
+                // videoPlayer.getVideoWidth().then(
+                //     (value) => { videoWidth = value; enableVideoHeroAutoresize(videoWrapperElement, videoPlayerWrapperElement, videoPlayer.element, videoWidth, videoHeight); },
+                //     (error) => { videoWidth = 800; enableVideoHeroAutoresize(videoWrapperElement, videoPlayerWrapperElement, videoPlayer.element, videoWidth, videoHeight); });
+                // videoPlayer.getVideoHeight().then(
+                //     (value) => { videoHeight = value; enableVideoHeroAutoresize(videoWrapperElement, videoPlayerWrapperElement, videoPlayer.element, videoWidth, videoHeight); },
+                //     (error) => { videoHeight = 450; enableVideoHeroAutoresize(videoWrapperElement, videoPlayerWrapperElement, videoPlayer.element, videoWidth, videoHeight); });
                 // API event calls to the Vimeo player can go here
                 videoPlayer.on('loaded', function() {
                     videoWrapperElement.removeAttribute('hidden');