+++ CHANGELOG.md
@@ -9,6 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- ### Change: Video Reveal Display & Autoplay
+  ### Changes to Video Reveal: 
+  Adds play/pause functionality to video reveal block videos, automatically toggling on click when image hides/shows. Fixes overlay to fit image.
+  
+  Resolves #388 
+---
+
 - ### Re-factor of Drupal Regions to more closely match D7 Express
   Added new regions to to .info
   Added new region rendering to page